Add Tour Detail Page

+++ src/main/java/com/travelagency/tirana/service/Impl/ReservationImpl/SaveReservationRequest.java
@@ -2,13 +2,12 @@ package com.travelagency.tirana.service.Impl.ReservationImpl;
 
 import lombok.Data;
 
-import java.time.LocalDateTime;
 import java.util.Date;
 
 @Data
 public class SaveReservationRequest {
     private long id;
-    private int tourId;
+    private long tourId;
     private String name;
     private String email;
     private String phoneNumber;
